Upgrade dependencies to latest versions (#11)
This change modernizes the package by upgrading the dependencies to the
latest version. The most notable (breaking) dependency upgrade is eslint
v9 which requires a new config format along with updating to a newer
foxglove eslint-plugin package. With this package we have more lint
errors that needed fixing along with the upgrade (the use of #private).
Doing this also requires we bump the es target in the tsconfig file so
typescript preserves the `#private` members rather than trying to use
polyfills like weakmap.