Indexed reading (#12)
Rather than reading all parsed messages into memory when a ULog file is
opened, this change scans the file and builds a message index. This
avoids having to hold the entire file in memory and allows ulog files
that would previously exceed the JS heap size.

The `readMessages` is updated to use this index to read and parse the
messages from the ulog file.

I've also added options to `readMessages` to filter to specific `msgIds`
and whether to include log messages or not (note log messages do not
have `msgIds`).